requirements.txt
asgiref==3.7.2
Django==4.2.4
environs==9.5.0
marshmallow==3.20.1
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3

djangorestframework~=3.14.0